Develop -> master (배포환경구성)

See merge request s08-webmobile2-sub2/S08P12C106!21

+++ osiopso/src/main/java/com/cojeans/osiopso/service/article/ArticleService.java
@@ -1,22 +1,22 @@
 package com.cojeans.osiopso.service.article;
 
-import com.cojeans.osiopso.dto.request.feed.ArticlePhotoRequestDto;
-import com.cojeans.osiopso.dto.request.feed.ArticleRequestDto;
-import com.cojeans.osiopso.dto.tag.TagDto;
+//import com.cojeans.osiopso.dto.request.feed.ArticleRequestDto;
+import com.cojeans.osiopso.dto.GapTimeVo;
+import com.cojeans.osiopso.entity.comment.Comment;
+import com.cojeans.osiopso.entity.feed.Advice;
 import com.cojeans.osiopso.entity.feed.Article;
-import com.cojeans.osiopso.entity.feed.ArticlePhoto;
-import com.cojeans.osiopso.entity.feed.ArticleTag;
-import com.cojeans.osiopso.entity.tag.Tag;
-import com.cojeans.osiopso.entity.user.User;
+        import com.cojeans.osiopso.entity.feed.ArticleTag;
+import com.cojeans.osiopso.entity.feed.Ootd;
 import com.cojeans.osiopso.repository.article.*;
-import com.cojeans.osiopso.repository.user.UserRepository;
-import com.cojeans.osiopso.security.UserDetail;
-import lombok.RequiredArgsConstructor;
+import com.cojeans.osiopso.repository.comment.CocommentRepository;
+import com.cojeans.osiopso.repository.comment.CommentLikeRepository;
+import com.cojeans.osiopso.repository.comment.CommentRepository;
+        import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.Date;
 import java.util.List;
-import java.util.Optional;
 
 @Service
 @Transactional(readOnly = false)
@@ -28,72 +28,49 @@ public class ArticleService {
     private final ArticlePhotoRepository articlePhotoRepository;
     private final TagRepository tagRepository;
     private final CommentRepository commentRepository;
-    private final UserRepository userRepository;
-
-
-    // 1. 게시물 먼저 저장
-// 2. Dto 필드 중, 태그의 개수만큼 태그 저장
-// 3. ArticleTag 에 게시물과 유저를 넣고 저장
-    public boolean createArticle(ArticleRequestDto articleRequestDto, Long id) {
-        // 게시물 저장
-        List<TagDto> tags = articleRequestDto.getTags();
-
-        Article article = articleRequestDto.toEntity(userRepository.getById(id), 0L);
-        Article savedArticle = articleRepository.save(article);
-
-
-        List<ArticlePhotoRequestDto> photos = articleRequestDto.getPhotos();
-
-        // 사진 저장
-        for (ArticlePhotoRequestDto photo : photos) {
-            articlePhotoRepository.save(ArticlePhoto.builder()
-                    .storeFilename(photo.getStoreFilename())
-                    .originFilename(photo.getOriginFilename())
-                    .article(savedArticle)
-                    .build());
-        }
-
-        if (tags != null) {
-            // 태그 저장
-            for (TagDto tag : tags) {
-                Tag tagE = tag.toEntity();
-                Tag tagSaved = tagRepository.save(tagE);
-
-                ArticleTag articleTagE = ArticleTag.builder()
-                        .article(article)
-                        .tag(tagSaved)
-                        .build();
-
-                articleTagRepository.save(articleTagE);
-            }
-        }
-
-        return true;
-    }
-
+    private final CommentLikeRepository commentLikeRepository;
+    private final CocommentRepository cocommentRepository;
+    private final ArticleLikeRepository articleLikeRepository;
 
     // 1번 article 을 지울 때..
-// 1번 article 을 외래키로 가진 article_tag 조회
-// 해당 article_tag 의 tag_id를 찾아서 tag 삭제
-// 해당 article_tag 삭제
-// 게시물 삭제
-    public boolean deleteArticle(Long articleNo) {
+    // 1번 article 을 외래키로 가진 article_tag 조회
+    // 해당 article_tag 의 tag_id를 찾아서 tag 삭제
+    // 해당 article_tag 삭제
+    // 게시물 삭제
+    public boolean deleteArticle(Long articleNo, Long userId) {
+        Article article = articleRepository.findById(articleNo).orElseThrow();
+
+        // 게시글 작성자만 삭제권한이 있다.
+        if (userId != article.getUser().getId()) {
+            return false;
+        }
+
         // 게시물과 관련된 태그들 삭제
         List<ArticleTag> articleTag = articleTagRepository.findByArticle_Id(articleNo);
         for (ArticleTag at : articleTag) {
             articleTagRepository.deleteById(at.getId());
-            tagRepository.deleteById(at.getTag().getId());
+//            tagRepository.deleteById(at.getTag().getId());
         }
 
         // 게시물과 관련된 사진삭제
         articlePhotoRepository.deleteByArticle_Id(articleNo);
         Long articleId = articleRepository.findById(articleNo).orElseThrow().getId();
 
+        // 게시물과 관련된 좋아요들 삭제
+        articleLikeRepository.deleteAllByArticle_Id(articleNo);
+
+        // 댓글과 관련된 좋아요들 삭제
+        commentLikeRepository.deleteAllByArticle_Id(articleNo);
+
+        // 게시물과 관련된 대댓글들 삭제
+        cocommentRepository.deleteAllByArticle_Id(articleNo);
+
         // 게시물과 관련된 댓글들 삭제
-        commentRepository.deleteByArticle_Id(articleNo);
+        commentRepository.deleteAllByArticle_Id(articleNo);
 
         // 게시물 삭제
         articleRepository.deleteById(articleId);
+        // gitpull 됐음
 
         // 확실히 지워진 경우 (삭제한 articleNo로 해당 게시물을 찾을 수 없어야 한다.)
         if (articleRepository.findById(articleNo).isEmpty()) {
@@ -101,5 +78,55 @@ public class ArticleService {
         } else {
             return false;
         }
+    }
+
+    public GapTimeVo getGapTime(Object object, Date date) {
+        Date createTime;
+
+        if (object instanceof Ootd) {
+            createTime = ((Ootd) object).getCreateTime();
+        } else if (object instanceof Advice) {
+            createTime = ((Advice) object).getCreateTime();
+        } else {
+            createTime = ((Comment) object).getCreateTime();
+        }
+
+        long createT = createTime.getTime();
+        long nowT = date.getTime();
+        long timeGap = (nowT - createT) / 1000;
+        float pastTime = timeGap / 1000;
+        String timeGapToString = "";
+
+        // l/1000 는 초 단위
+        if (timeGap < 60) {
+            timeGapToString = Long.toString(timeGap) + "s";
+        } else if (timeGap < 3600) { // 60초 ~ 3600초(1분 ~ 60분) 는 분 단위
+            timeGapToString = Long.toString(timeGap / 60) + "m";
+        } else if (timeGap < 84000) { // 3601초 ~ 84000초 (1시간 ~ 24시간) 는 시간 단위
+            timeGapToString = Long.toString(timeGap / 3600) + "h";
+        } else if (timeGap < 2520000) { // 84001초 ~  (1일 ~ 30일) 는 일단위
+            timeGapToString = Long.toString(timeGap / 84000) + "d";
+        }
+
+        return GapTimeVo.builder()
+                .pastTime(pastTime)
+                .timeGapToString(timeGapToString)
+                .build();
+    }
+
+    public void reportArticle(Long articleNo) {
+        Article article = articleRepository.findById(articleNo).orElseThrow();
+
+        articleRepository.save(Article.builder()
+                .id(articleNo)
+                .dtype(article.getDtype())
+                .content(article.getContent())
+                .dtype(article.getDtype())
+                .user(article.getUser())
+                .report(article.getReport() + 1)
+                .createTime(article.getCreateTime())
+                .modifyTime(article.getModifyTime())
+                .hit(article.getHit())
+                .build());
     }
 }
